feat: creo archivo theme.ts para separar responsabilidad
